[ramda] split tests to separate modules: 'join' - 'mean' (#38570)
* move 'join' tests to separate file

* move 'juxt' tests to separate file

* move 'keys' tests to separate file

* move 'keysIn' tests to separate file

* move 'last' tests to separate file

* move 'lastIndexOf' tests to separate file

* move 'length' tests to separate file

* move 'lens' tests to separate file

* move 'lensIndex' tests to separate file

* move 'lensProp' tests to separate file

* move 'lt' tests to separate file

* move 'lte' tests to separate file

* move 'map' tests to separate file

* move 'mapAccum' tests to separate file

* move 'mapAccumRight' tests to separate file

* move 'mapObjIndexed' tests to separate file

* move 'mathMod' tests to separate file

* move 'max' tests to separate file

* move 'maxBy' tests to separate file

* move 'mean' tests to separate file
