+++ types/ramda/test/keysIn-tests.ts
@@ -1,0 +1,12 @@
+import * as R from 'ramda';
+
+class F {
+  [k: string]: string;
+  x = 'X';
+  y = 'Y';
+}
+
+() => {
+  const f = new F();
+  R.keysIn(f); // => ['x', 'y']
+};
