+++ types/ramda/test/last-tests.ts
@@ -1,0 +1,7 @@
+import * as R from 'ramda';
+
+() => {
+  R.last(['fi', 'fo', 'fum']); // => 'fum'
+  R.last('abc'); // => 'c'
+  R.last(''); // => ''
+};
