+++ types/ramda/test/length-tests.ts
@@ -1,0 +1,6 @@
+import * as R from 'ramda';
+
+() => {
+  R.length([]); // => 0
+  R.length([1, 2, 3]); // => 3
+};
