+++ types/ramda/test/mathMod-tests.ts
@@ -1,0 +1,13 @@
+import * as R from 'ramda';
+
+() => {
+  R.mathMod(-17, 5); // => 3
+  R.mathMod(17, 5); // => 2
+  R.mathMod(17, -5); // => NaN
+  R.mathMod(17, 0); // => NaN
+  R.mathMod(17.2, 5); // => NaN
+  R.mathMod(17, 5.3); // => NaN
+
+  const seventeenMod = R.mathMod(17);
+  seventeenMod(3); // => 2
+};
